split two folders e2e and unit
