+++ package.json
@@ -4,18 +4,18 @@
   "description": "Shows Wikipedia article preview in a popup",
   "main": "dist/wikipedia-previews.production.js",
   "scripts": {
-    "test": "mocha --require @babel/register",
-    "test:watch": "mocha --require @babel/register -w",
+    "test": "mocha test/unit --require @babel/register",
+    "test:watch": "mocha test/unit --require @babel/register -w",
     "coverage": "nyc npm test",
     "build:dev": "webpack --mode development",
     "build:prod": "webpack --mode production",
     "build": "npm run build:dev && npm run build:prod",
     "dev": "webpack-dev-server --mode development",
-    "selenium-safari": "wdio test/config/wdio.safari.conf.js",
-    "selenium-edge": "wdio test/config/wdio.edge.conf.js",
-    "selenium-chrome": "wdio test/config/wdio.chrome.conf.js",
-    "selenium-ie": "wdio test/config/wdio.ie.conf.js",
-    "selenium-firefox": "wdio test/config/wdio.ff.conf.js"
+    "selenium-safari": "wdio test/e2e/config/wdio.safari.conf.js",
+    "selenium-edge": "wdio test/e2e/config/wdio.edge.conf.js",
+    "selenium-chrome": "wdio test/e2e/config/wdio.chrome.conf.js",
+    "selenium-ie": "wdio test/e2e/config/wdio.ie.conf.js",
+    "selenium-firefox": "wdio test/e2e/config/wdio.ff.conf.js"
   },
   "author": "Wikimedia Foundation",
   "license": "MIT",
