+++ test/unit/cachedFetch.test.js
@@ -1,5 +1,5 @@
 const assert = require('assert')
-const { cachedFetch } = require('../src/cachedFetch')
+const { cachedFetch } = require('../../src/cachedFetch')
 
 const fetchMock = (data) => {
 	return () => {
